+++ src/backends/sentry_backend_inproc.c
@@ -595,13 +595,16 @@ handle_ucontext(const sentry_ucontext_t *uctx)
             sentry__envelope_add_session(envelope, session);
 
             if (options->attach_screenshot) {
-                sentry_path_t *screenshot_path
-                    = sentry__screenshot_get_path(options);
-                if (sentry__screenshot_capture(screenshot_path)) {
+                size_t count
+                    = sentry__screenshot_capture(options->run->run_path);
+                for (size_t i = 0; i < count; i++) {
+                    sentry_path_t *screenshot_path
+                        = sentry__screenshot_get_path(
+                            options->run->run_path, i);
                     sentry__envelope_add_attachment(
                         envelope, screenshot_path, NULL);
+                    sentry__path_free(screenshot_path);
                 }
-                sentry__path_free(screenshot_path);
             }
 
             // capture the envelope with the disk transport